+++ src/draco/io/parser_utils.cc
@@ -25,7 +25,7 @@ namespace parser {
 void SkipCharacters(DecoderBuffer *buffer, const char *skip_chars) {
   if (skip_chars == nullptr)
     return;
-  const int num_skip_chars = strlen(skip_chars);
+  const int num_skip_chars = (int)strlen(skip_chars);
   char c;
   while (buffer->Peek(&c)) {
     // Check all characters in the pattern.
@@ -132,7 +132,7 @@ bool ParseFloat(DecoderBuffer *buffer, float *value) {
     }
   }
 
-  *value = (sign < 0) ? -v : v;
+  *value = (sign < 0) ? (float)-v : (float)v;
   return true;
 }
