+++ src/draco/mesh/mesh.h
@@ -50,7 +50,7 @@ class Mesh : public PointCloud {
   void AddFace(const Face &face) { faces_.push_back(face); }
 
   void SetFace(FaceIndex face_id, const Face &face) {
-    if (face_id >= faces_.size()) {
+    if (face_id >= (uint32_t)faces_.size()) {
       faces_.resize(face_id.value() + 1, Face());
     }
     faces_[face_id] = face;
@@ -60,7 +60,7 @@ class Mesh : public PointCloud {
   // existing ones if necessary.
   void SetNumFaces(size_t num_faces) { faces_.resize(num_faces, Face()); }
 
-  FaceIndex::ValueType num_faces() const { return faces_.size(); }
+  FaceIndex::ValueType num_faces() const { return (uint32_t)faces_.size(); }
   const Face &face(FaceIndex face_id) const {
     DRACO_DCHECK_LE(0, face_id.value());
     DRACO_DCHECK_LT(face_id.value(), static_cast<int>(faces_.size()));
@@ -135,7 +135,7 @@ struct MeshHasher {
     PointCloudHasher pc_hasher;
     size_t hash = pc_hasher(mesh);
     // Hash faces.
-    for (FaceIndex i(0); i < mesh.faces_.size(); ++i) {
+    for (FaceIndex i(0); i < (uint32_t)mesh.faces_.size(); ++i) {
       for (int j = 0; j < 3; ++j) {
         hash = HashCombine(mesh.faces_[i][j].value(), hash);
       }
